requirements.txt
google-genai
python-dotenv
mysql-connector-python
imap-tools
smtplib
youtube-transcript-api
requests
